requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langgraph>=0.0.21
beautifulsoup4>=4.12.2
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
chromadb>=0.4.22
langchain-chroma>=0.0.1
fake-useragent>=1.3.0
streamlit>=1.31.0
streamlit-chat>=0.1.1
transformers>=4.38.1
sentence-transformers>=2.3.1
tqdm>=4.66.1
tavily-python>=0.2.8
langchain-openai>=0.0.2
python-json-logger>=2.0.7
colorlog>=6.7.0 